Refactored World Gen Config
