+++ src/main/java/epicsquid/roots/world/WorldGenHut.java
@@ -1,21 +1,14 @@
 package epicsquid.roots.world;
 
-import java.util.Random;
-
 import epicsquid.mysticallib.network.MessageTEUpdate;
 import epicsquid.mysticallib.network.PacketHandler;
 import epicsquid.mysticallib.util.NoiseGenUtil;
 import epicsquid.mysticallib.util.Util;
-import epicsquid.roots.ConfigManager;
-import epicsquid.roots.init.ModBlocks;
-import net.minecraft.block.BlockChest;
-import net.minecraft.block.BlockFlowerPot;
+import epicsquid.roots.config.WorldGenConfig;
 import net.minecraft.block.BlockGrass;
-import net.minecraft.block.BlockLog;
 import net.minecraft.block.state.IBlockState;
 import net.minecraft.init.Blocks;
 import net.minecraft.tileentity.TileEntityChest;
-import net.minecraft.util.EnumFacing;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.DimensionType;
@@ -23,6 +16,8 @@ import net.minecraft.world.World;
 import net.minecraft.world.chunk.IChunkProvider;
 import net.minecraft.world.gen.IChunkGenerator;
 import net.minecraftforge.fml.common.IWorldGenerator;
+
+import java.util.Random;
 
 public class WorldGenHut extends StructureBase implements IWorldGenerator {
   public WorldGenHut() {
@@ -62,13 +57,13 @@ public class WorldGenHut extends StructureBase implements IWorldGenerator {
 
   @Override
   public void generate(Random random, int chunkX, int chunkZ, World world, IChunkGenerator chunkGenerator, IChunkProvider chunkProvider) {
-    if (ConfigManager.worldGen.hutChance <= 0) {
+    if (WorldGenConfig.worldGen.hutChance <= 0) {
       return;
     }
     if (world.provider.getDimension() == DimensionType.OVERWORLD.getId() && !world.isRemote) {
       int xx = chunkX * 16 + 12 + Util.rand.nextInt(8);
       int zz = chunkZ * 16 + 12 + Util.rand.nextInt(8);
-      if (random.nextInt(ConfigManager.worldGen.hutChance) == 0) {
+      if (random.nextInt(WorldGenConfig.worldGen.hutChance) == 0) {
         int height = world.getHeight(xx, zz) - 1;
         if (height > 0 && world.getBlockState(new BlockPos(xx, height, zz)).getBlock() instanceof BlockGrass) {
           boolean canGenerate = true;
